+++ tests/models/test_models.py
@@ -396,7 +396,6 @@ def test_metric_loss_metadata_routing(estimator, dataset):
     cost_loss = Metric(cost_matrix, Cost())
 
     estimator = set_metric_loss(clone(estimator), cost_loss)
-    estimator._append_params_to_metadata_routing()
 
     with config_context(enable_metadata_routing=True):
         if isinstance(estimator, CSThresholdClassifier | CSRateClassifier):